# openapi/admin.openapi.yaml
openapi: 3.1.0
info:
  title: Whistlelads Faka Cards & Files API
  description: |
    面向脚本/自动化的管理接口：**直接用后台密码**操作卡密与文件。
    无需登录、无需 Session、不含 Spaces。

    ## 鉴权
    每个请求带上后台密码（任选其一）：

    | 方式 | 示例 |
    |------|------|
    | Header `X-Admin-Password`（推荐） | `X-Admin-Password: your-password` |
    | Header `Authorization: Bearer` | `Authorization: Bearer your-password` |
    | Header `Authorization: Password` | `Authorization: Password your-password` |

    密码错误或缺失 → `401`，body：`{"ok":false,"error":"未登录"}`

    ## 空间
    所有接口默认操作 **默认空间**（id 最小）。本规范不暴露 Spaces API。

    ## 常见流程
    1. 上传库存文件：`POST /api/admin/files/upload`
    2. 生成卡密：`POST /api/admin/cards`（`file_count` + `quantity`）
    3. 下载卡密并标为待使用：`POST /api/admin/cards/download`（`mark_pending: true`）
    4. 改状态：`POST /api/admin/cards/status` / `POST /api/admin/files/status`
  version: 1.1.0
servers:
  - url: https://key.whistlelads.com
    description: Production
  - url: http://127.0.0.1:18743
    description: Local
...
security:
  - AdminPassword: []
paths:
  /api/admin/cards:
    get:
      tags: [Cards]
      summary: 分页列出卡密
      operationId: listCards
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: q
          in: query
          description: 卡密模糊搜索
          schema:
            type: string
        - name: status
          in: query
          description: 状态过滤
          schema:
            $ref: '#/components/schemas/CardStatus'
        - name: start
          in: query
          description: used_at 起始 `YYYY-MM-DD`
          schema:
            type: string
            format: date
        - name: end
          in: query
          description: used_at 结束 `YYYY-MM-DD`
          schema:
            type: string
            format: date
      responses:
        '200':
          description: 卡密列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Cards]
      summary: 批量生成卡密
      operationId: createCards
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCardsRequest'
            examples:
              oneEach:
                summary: 生成 100 张，每张 1 个文件
                value:
                  file_count: 1
                  quantity: 100
      responses:
        '200':
          description: 生成成功，直接返回新建卡密的 ids 与 codes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateCardsResponse'
              example:
                ok: true
                message: 已生成 2 张卡密
                count: 2
                ids: [101, 102]
                codes:
                  - DEFAULT-ABCDEFGHIJKLMNOPQRSTUVWXYZ012345
                  - DEFAULT-ZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZ
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    AdminPassword:
      type: apiKey
      in: header
      name: X-Admin-Password
      description: |
        后台密码（TIKAWANG_AUTH_PASSWORD）。
        也可用 `Authorization: Bearer <password>`。
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: page_size
      in: query
      description: 50/100/200/500/1000/2000/5000/10000，默认 200
      schema:
        type: integer
        enum: [50, 100, 200, 500, 1000, 2000, 5000, 10000]
        default: 200
  responses:
    Unauthorized:
      description: 密码错误或未提供
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
          example:
            ok: false
            error: 未登录
    Error:
      description: 参数/业务错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
  schemas:
...
    ErrorBody:
      type: object
      required: [ok, error]
      properties:
        ok: { type: boolean, const: false }
        error: { type: string }
    CardStatus:
      type: string
      enum: [available, pending, sold, voided]
      description: |
        available=新生成 · pending=待兑换 · sold=已用 · voided=作废
...
    Pagination:
      type: object
      required: [page, page_size, total, total_pages, has_prev, has_next, prev_page, next_page]
      properties:
        page: { type: integer }
        page_size: { type: integer }
        total: { type: integer }
        total_pages: { type: integer }
        has_prev: { type: boolean }
        has_next: { type: boolean }
        prev_page: { type: integer }
        next_page: { type: integer }
    Space:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        card_prefix: { type: string }
    Card:
      type: object
      required: [id, code, file_count, status, created_at]
      properties:
        id: { type: integer }
        code:
          type: string
          example: DEFAULT-ABCDEFGHIJKLMNOPQRSTUVWXYZ012345
        file_count:
          type: integer
          minimum: 1
        status:
          $ref: '#/components/schemas/CardStatus'
        created_at: { type: string, example: "2026-07-18 01:00:00" }
        used_at: { type: string }
        voided_at: { type: string }
...
    CardListResponse:
      type: object
      required: [ok, cards, pagination]
      properties:
        ok: { type: boolean, const: true }
        cards:
          type: array
          items:
            $ref: '#/components/schemas/Card'
        pagination:
          $ref: '#/components/schemas/Pagination'
        current_space:
          $ref: '#/components/schemas/Space'
...
    CreateCardsRequest:
      type: object
      required: [file_count]
      properties:
        file_count:
          type: integer
          minimum: 1
          maximum: 10000
          description: 每张卡绑定文件数
          example: 1
        quantity:
          type: integer
          minimum: 1
          maximum: 10000
          default: 1
          description: 生成张数
          example: 100
    CreateCardsResponse:
      type: object
      required: [ok, message, count, ids, codes]
      properties:
        ok:
          type: boolean
          const: true
        message:
          type: string
          example: 已生成 2 张卡密
        count:
          type: integer
          description: 生成张数
          example: 2
        ids:
          type: array
          description: 新建卡密数据库 ID（顺序与 codes 对应）
          items:
            type: integer
            minimum: 1
          example: [101, 102]
        codes:
          type: array
          description: 新建卡密明文（顺序与 ids 对应）
          items:
            type: string
          example:
            - DEFAULT-ABCDEFGHIJKLMNOPQRSTUVWXYZ012345
            - DEFAULT-ZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZZ
...